* configure, configure.ac: changed warning into notice * sdas/asz80/z80mch.c: cosmetic * src/sdcc.vcxproj, * src/sdcc.vcxproj.filters: added missing SDCC.lex & SDCC.y * support/cpp/libcpp/files.c, * support/cpp/libcpp/identifiers.c, * support/cpp/libcpp/init.c, * support/cpp/libcpp/symtab.c: fixed warnings for unsigned size_t * support/regression/tests/bitopcse.c: only use idata for mcs51 * support/regression/tests/gcc-torture-execute-20021120-1.c: disabled for   hc08 & s08, uses way too much spillocs in DSEG * support/regression/tests/longlong.c: split the test in 3 parts * support/sdbinutils/bfd/targets.c: added "ASXXXX" * support/sdbinutils/libiberty/simple-object-elf.c: fixed warning
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@10764 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/support/cpp/libcpp/files.c
@@ -1179,7 +1179,7 @@ _cpp_init_files (cpp_reader *pfile)
 						    nonexistent_file_hash_eq,
 						    NULL, xcalloc, free);
   _obstack_begin (&pfile->nonexistent_file_ob, 0, 0,
-		  (void *(*) (long)) xmalloc,
+		  (void *(*) (size_t)) xmalloc,
 		  (void (*) (void *)) free);
 }
